feat: number input, formerly currency input
