+++ packages/next-common/lib/input/number/index.d.js
@@ -1,0 +1,10 @@
+/**
+ * @typedef {{
+ * allowDecimals?: boolean
+ * controls?: boolean
+ * keyboard?: boolean
+ * step?: number | string
+ * min?: number | string
+ * max?: number | string
+ * } & InputProps} NumberInputProps
+ */
